MeteringPointTypeMapper
